pyproject.toml
[project]
name = "aoa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.3",
    "smolagents[litellm,mcp]>=0.2.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.6.2",
    "langchain-anthropic>=0.3.18",
    "langchain>=0.3.27",
    "fastmcp>=2.11.0",
    "a2a-sdk[sqlite]>=0.3.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.7",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_return_any = false
check_untyped_defs = true
